move Filter component out
