+++ assets/js/common/Filter/index.js
@@ -1,0 +1,3 @@
+import Filter from './Filter';
+
+export default Filter;
